[IMP] tools/clean-matt-databases.sh: allow to specify another db prefix

+++ tools/clean-matt-databases.sh
@@ -2,7 +2,12 @@
 set -euo pipefail
 T=$(mktemp)
 trap 'rm -f "$T"' EXIT
-psql -d postgres -tAF" " -c '\l matt-*' | awk '{if($2 == ENVIRON["USER"]) print $1}' > "$T"
+: "${PREFIX:=matt-}"
+psql -d postgres -tAF" " -c "\l ${PREFIX}*" | awk '{if($2 == ENVIRON["USER"]) print $1}' > "$T"
 S=$(wc -l "$T" | awk '{print $1}')
-echo "Deleting $S databases"
+if [[ $S -eq 0 ]]; then
+    echo "🥳 No databases to remove"
+    exit 0
+fi
+echo "🗑  Deleting $S databases"
 (while read -r db; do echo "$db"; dropdb "$db"; done < "$T") | pv -betlap -s "$S" >/dev/null
